# .github/workflows/openAPI.yml
openapi: '3.0.2'
info:
  title: My Wedding Day
  description: Alta Project 
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://app.swaggerhub.com/apis/Nathannov24/WeddingDay/1.0
...
paths:
...
  /package:
    post:
      tags:
        - Package
      summary: Post a Package to Database
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestPostPackage'
            examples:
              userData1:
                description: "Example data Package"
                value: {
                  PackageName: "Nikah Muda Asoy",
                  price: "270000000",
                  pax: "1000",
                  PackageDesc: "ini cuman package desc kok",
                  urlphoto: "https://trello.com/1/cards/61b2121947c45d456109ac16/attachments/61b34d5757f4352a736dff8c/previews/61b34d5857f4352a736e000a/download/wedding_%282%29.jpg"
                }
      responses:
        '201':
          description: Success Response
          content:
            application/json:
              examples:
                responseData1:
                  description: response success 1
                  value: {
                    message: "Success create new package",
                    status: "success"
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: failed
                  value: {
                    message: "data is already exist",
                    status: "failed"
                  }
    get:
      tags:
        - Package
      summary: Get All Package available in Database
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              examples:
                userData1:
                  description: "Example data package"
                  value: {
                    data: [
                      {
                        PackageID: 1,
                        ProfileWoID: 1,
                        PackageName: "Nikah Muda Asoy",
                        price: "270000000",
                        pax: "1000",
                        PackageDesc: "ini cuman package desc kok",
                        photo: "https://trello.com/1/cards/61b2121947c45d456109ac16/attachments/61b34d5757f4352a736dff8c/previews/61b34d5857f4352a736e000a/download/wedding_%282%29.jpg"
                      },
                      {
                        PackageID: 2,
                        ProfileWoID: 2,
                        PackageName: "Nikah Muda Asoy euy",
                        price: "280000000",
                        pax: "1000",
                        PackageDesc: "ini cuman package desc kok",
                        photo: "https://trello.com/1/cards/61b2121947c45d456109ac16/attachments/61b34d5757f4352a736dff8c/previews/61b34d5857f4352a736e000a/download/wedding_%282%29.jpg"
                      }
                    ],
                    message: "success get package",
                    status: "success"
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: failed
                  value: {
                    message: "failed get package",
                    status: "failed"
                  }
...
  /package/:id:
    get:
      tags:
        - Package
      summary: Get Package by ID
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              examples:
                userData1:
                  description: "Example data package"
                  value: {
                    data: {
                        PackageID: 1,
                        ProfileWoID: 1,
                        PackageName: "Nikah Muda Asoy",
                        price: "270000000",
                        pax: "1000",
                        PackageDesc: "ini cuman package desc kok",
                        photo: "https://trello.com/1/cards/61b2121947c45d456109ac16/attachments/61b34d5757f4352a736dff8c/previews/61b34d5857f4352a736e000a/download/wedding_%282%29.jpg"
                      },
                    message: "success get package",
                    status: "success"
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: failed
                  value: {
                    message: "failed get package",
                    status: "failed"
                  }
    put:
      tags:
        - Package
      summary: Update Package
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestPackage'
            examples:
              userData1:
                description: "Example data homestay"
                value: {
                  PackageName: "Nikah Muda Asoy",
                  price: "270000000",
                  pax: "1000",
                  PackageDesc: "ini cuman package desc kok",
                  photo: "https://trello.com/1/cards/61b2121947c45d456109ac16/attachments/61b34d5757f4352a736dff8c/previews/61b34d5857f4352a736e000a/download/wedding_%282%29.jpg"
                }
      responses:
        '201':
          description: Success Response
          content:
            application/json:
              examples:
                responseData1:
                  description: success
                  value: {
                    message: "Success edit package",
                    status: "success"
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: missing jwt (token)
                  value: {
                    message: "missing or malformed jwt"
                  }
    delete:
      tags:
        - Package
      summary: Delete Package By ID
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              examples:
                userData1:
                  description: success
                  value: {
                    message: "success delete package",
                    status: "success"
                  }
        '401':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData1:
                  description: failed
                  value: {
                    message: "Unauthorized Access",
                    status: "Unauthorized"
                  }
  
####################################
#   RESERVATIONS
####################################
...
components:
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  ####################################
  #   SCHEMAS
  ####################################
  schemas:
...
    requestPostPackage:
      type: object
      properties: {
        "PackageName":{
          type: string
        },
        "price":{
          type: integer
        },
        "pax":{
          type: integer
        },
        "PackageDesc":{
          type: string
        },
        "urlphoto":{
          type: string
        },
      }
...
    requestPackage:
      type: object
      properties: {
        "data": {
          type: array,
          items: {
            type: object
          },
        }
      }
...